requirements.txt
alembic==1.4.2
click==7.1.1
Flask==1.1.2
Flask-Login==0.5.0
flask-marshmallow==0.11.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.1
psycopg2==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.16
Werkzeug==1.0.1
WTForms==2.2.1
